+++ src/ImageProcessorCore/Samplers/Resamplers/HermiteResampler.cs
@@ -13,13 +13,13 @@ namespace ImageProcessorCore.Processors
     public class HermiteResampler : IResampler
     {
         /// <inheritdoc/>
-        public double Radius => 2;
+        public float Radius => 2;
 
         /// <inheritdoc/>
-        public double GetValue(double x)
+        public float GetValue(float x)
         {
-            const double B = 0;
-            const double C = 0;
+            const float B = 0F;
+            const float C = 0F;
 
             return ImageMaths.GetBcValue(x, B, C);
         }
