+++ src/ImageProcessorCore/Samplers/Resamplers/Lanczos5Resampler.cs
@@ -13,22 +13,22 @@ namespace ImageProcessorCore
     public class Lanczos5Resampler : IResampler
     {
         /// <inheritdoc/>
-        public double Radius => 5;
+        public float Radius => 5;
 
         /// <inheritdoc/>
-        public double GetValue(double x)
+        public float GetValue(float x)
         {
-            if (x < 0)
+            if (x < 0F)
             {
                 x = -x;
             }
 
-            if (x < 5)
+            if (x < 5F)
             {
-                return ImageMaths.SinC(x) * ImageMaths.SinC(x / 5d);
+                return ImageMaths.SinC(x) * ImageMaths.SinC(x / 5F);
             }
 
-            return 0;
+            return 0F;
         }
     }
 }
